requirements.txt
pandas
pdfplumber
openpyxl
google-cloud-translate==2.0.1